refactor(platform): extract websocket upgrade handler
